test(subroutines): drop the per-chart template render tests
